+++ crates/axiom-persist/src/loader.rs
@@ -85,7 +85,7 @@ fn state_to_snapshot(state: &StoredEngineState) -> axiom_runtime::EngineSnapshot
     use axiom_config::DomainConfig;
 
     let domains = state.domains.iter().map(|sd| DomainSnapshot {
-        domain_id:   sd.domain_id,
+        domain_id:   sd.domain_id as u16,
         config:      sd.config.unwrap_or_else(|| DomainConfig::factory_void(sd.domain_id as u16, 0)),
         tokens:      sd.tokens.clone(),
         connections: sd.connections.clone(),
